refactoring editor tests and docs
