+++ Assets/Tests/EditMode/GnssSensorTest.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b96707fd986b4d610b05c0f27bd3c12a
+guid: 4ed8f29bb8d9f17458f48f7e5638bd59
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
