+++ Assets/Tests/EditMode/ImuSensorTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1c5c4e1de272b1f409f784cb9b623042
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
